# openapi/openapi.v1.yaml
openapi: 3.1.0
info:
  title: Energy Allocation API
  version: 1.0.0
  description: >-
    Tenant-aware API for onboarding facilities, ingesting data, generating model-driven
    recommendations, and tracking realized savings.
servers:
  - url: https://api.energyallocation.com
security:
  - oauth2: [client_admin, facility_manager, energy_analyst]
paths:
...
  /v1/facilities/{facility_id}/features/materialize:
    post:
      tags: [Modeling]
      summary: Trigger feature materialization
      operationId: materializeFeatures
      parameters:
        - $ref: '#/components/parameters/facilityId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MaterializeFeaturesRequest'
      responses:
        '202':
          description: Materialization queued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AsyncJobResponse'
  /v1/facilities/{facility_id}/models/runs:
    post:
      tags: [Modeling]
      summary: Start model run
      operationId: startModelRun
      parameters:
        - $ref: '#/components/parameters/facilityId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateModelRunRequest'
      responses:
        '202':
          description: Model run queued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ModelRunQueuedResponse'
  /v1/models/runs/{model_run_id}:
    get:
      tags: [Modeling]
      summary: Get model run details
      operationId: getModelRunDetail
      parameters:
        - $ref: '#/components/parameters/modelRunId'
      responses:
        '200':
          description: Model run detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ModelRunDetail'
...
  /v1/facilities/{facility_id}/savings:
    get:
      tags: [Monitoring]
      summary: Get savings summary
      operationId: getSavings
      parameters:
        - $ref: '#/components/parameters/facilityId'
        - in: query
          name: from
          schema: { type: string, format: date }
        - in: query
          name: to
          schema: { type: string, format: date }
      responses:
        '200':
          description: Savings summary
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SavingsSummary'
  /v1/facilities/{facility_id}/alerts:
    get:
      tags: [Monitoring]
      summary: List alerts
      operationId: listAlerts
      parameters:
        - $ref: '#/components/parameters/facilityId'
        - $ref: '#/components/parameters/status'
        - $ref: '#/components/parameters/severity'
        - $ref: '#/components/parameters/limit'
        - $ref: '#/components/parameters/cursor'
      responses:
        '200':
          description: Alert list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AlertListResponse'
  /v1/facilities/{facility_id}/drift-events:
    get:
      tags: [Monitoring]
      summary: List drift events
      operationId: listDriftEvents
      parameters:
        - $ref: '#/components/parameters/facilityId'
        - $ref: '#/components/parameters/limit'
        - $ref: '#/components/parameters/cursor'
      responses:
        '200':
          description: Drift event list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DriftEventListResponse'
  /v1/facilities/{facility_id}/retraining-jobs:
    get:
      tags: [Monitoring]
      summary: List retraining jobs
      operationId: listRetrainingJobs
      parameters:
        - $ref: '#/components/parameters/facilityId'
        - $ref: '#/components/parameters/status'
        - $ref: '#/components/parameters/limit'
        - $ref: '#/components/parameters/cursor'
      responses:
        '200':
          description: Retraining jobs list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RetrainingJobListResponse'
  /v1/alerts/{alert_id}/acknowledge:
    post:
      tags: [Monitoring]
      summary: Acknowledge an alert
      operationId: acknowledgeAlert
      parameters:
        - in: path
          name: alert_id
          required: true
          schema: { type: string }
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                note: { type: string }
      responses:
        '200':
          description: Alert acknowledged
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Alert'
  /v1/alerts/{alert_id}/incident:
    post:
      tags: [Monitoring]
      summary: Open incident from alert
      operationId: openIncidentFromAlert
      parameters:
        - in: path
          name: alert_id
          required: true
          schema: { type: string }
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                note: { type: string }
      responses:
        '200':
          description: Incident opened
          content:
            application/json:
              schema:
                type: object
                properties:
                  incident_id: { type: string }
                  alert_id: { type: string }
                  status: { type: string }
...
components:
  securitySchemes:
    oauth2:
      type: oauth2
      flows:
        clientCredentials:
          tokenUrl: https://auth.energyallocation.com/oauth/token
          scopes:
            client_admin: Manage client-level resources
            facility_manager: Manage facility operations
            energy_analyst: Read analytics and model outputs
            viewer: Read-only access
            ops_admin: Internal operations access
            ml_engineer: Internal ML administration
  parameters:
...
    facilityId: { in: path, name: facility_id, required: true, schema: { type: string } }
...
    modelRunId: { in: path, name: model_run_id, required: true, schema: { type: string } }
...
    status: { in: query, name: status, schema: { type: string } }
    severity: { in: query, name: severity, schema: { type: string, enum: [info, warning, high, critical] } }
...
    limit: { in: query, name: limit, schema: { type: integer, default: 50, minimum: 1, maximum: 200 } }
    cursor: { in: query, name: cursor, schema: { type: string } }
...
  schemas:
...
    MaterializeFeaturesRequest:
      type: object
      properties:
        window_start: { type: string, format: date-time }
        window_end: { type: string, format: date-time }
        granularity: { type: string, enum: [minute, hour, day] }
      required: [window_start, window_end, granularity]
    CreateModelRunRequest:
      type: object
      properties:
        feature_set_version: { type: string }
        model_family: { type: string }
        objective: { type: string }
        explainability: { type: boolean }
      required: [feature_set_version, model_family, objective]
    ModelRunQueuedResponse:
      type: object
      properties:
        model_run_id: { type: string }
        status: { type: string }
        feature_rows: { type: integer }
        model_version: { type: string }
      required: [model_run_id, status]
    ModelRunDetail:
      type: object
      properties:
        model_run_id: { type: string }
        model_version: { type: string }
        training_window:
          type: object
          properties:
            start: { type: string, format: date-time }
            end: { type: string, format: date-time }
        validation:
          type: object
          properties:
            train_mae_kw: { type: number }
            train_mape: { type: number }
            holdout_mae_kw: { type: number }
            holdout_mape: { type: number }
            training_rows: { type: number }
        selected_hyperparameters:
          type: object
          additionalProperties: true
        drift_baseline_id: { type: string }
        shap_artifacts:
          type: array
          items: { type: string, format: uri }
        anomaly_summary:
          type: object
          additionalProperties: true
        model_artifact_path: { type: string }
        retraining_triggered: { type: boolean }
        retraining_job_id: { type: [string, 'null'] }
        drift_event_id: { type: [string, 'null'] }
        retraining_threshold: { type: number }
        recommendations_generated: { type: integer }
      required: [model_run_id, model_version]
...
    SavingsSummary:
      type: object
      properties:
        projected_savings_eur: { type: number }
        realized_savings_eur: { type: number }
        realization_rate: { type: number }
        by_appliance:
          type: array
          items:
            type: object
            properties:
              appliance_id: { type: string }
              projected_eur: { type: number }
              realized_eur: { type: number }
            required: [appliance_id, projected_eur, realized_eur]
      required: [projected_savings_eur, realized_savings_eur, realization_rate]
    Alert:
      type: object
      properties:
        alert_id: { type: string }
        source_type: { type: string }
        source_id: { type: string }
        facility_id: { type: string }
        client_id: { type: string }
        severity: { type: string, enum: [info, warning, high, critical] }
        state: { type: string, enum: [open, acknowledged, investigating, resolved, closed, suppressed] }
        owner_role: { type: string }
        routing_policy: { type: string }
        sla_minutes: { type: integer }
        dedupe_key: { type: string }
      required: [alert_id, source_type, severity, state, facility_id, client_id]
    AlertListResponse:
      type: object
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/Alert' }
        next_cursor: { type: [string, 'null'] }
      required: [items]
    DriftEvent:
      type: object
      properties:
        drift_event_id: { type: string }
        feature_distribution_change: { type: string }
        monitored_segment: { type: string }
        threshold_breached: { type: number }
        affected_model: { type: string }
        retraining_action: { type: string }
      required: [drift_event_id, feature_distribution_change, monitored_segment, affected_model]
    DriftEventListResponse:
      type: object
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/DriftEvent' }
        next_cursor: { type: [string, 'null'] }
      required: [items]
    RetrainingJob:
      type: object
      properties:
        job_id: { type: string }
        facility_id: { type: string }
        model_run_id: { type: string }
        status: { type: string, enum: [queued, running, completed, failed] }
        reason: { type: string }
        threshold: { type: number }
        observed_holdout_mape: { type: number }
      required: [job_id, facility_id, status]
    RetrainingJobListResponse:
      type: object
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/RetrainingJob' }
        next_cursor: { type: [string, 'null'] }
      required: [items]
    AsyncJobResponse:
      type: object
      properties:
        job_id: { type: string }
        status: { type: string, enum: [queued, running, completed, failed] }
      required: [job_id, status]
...